requirements.txt
python-telegram-bot>=20.0
httpx>=0.23.0
dnspython>=2.3.0
python-dotenv>=0.21.0
requests